# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API server for asset management system.",
        "title": "Asset Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:6500",
    "paths": {
...
        "/sys/department/list": {
            "post": {
                "description": "List Departments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "List Departments",
                "parameters": [
                    {
                        "description": "List Action Record Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DepartmentPageDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DepartmentList"
                        }
                    }
                }
            }
        },
        "/sys/department/{id}": {
            "get": {
                "description": "Get one Department record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Get one Department record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    }
                }
            }
        },
        "/sys/dpartment/create": {
            "post": {
                "description": "Create one dpartment record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Create one dpartment record",
                "parameters": [
                    {
                        "description": "Department Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralCreateResponseBody"
                        }
                    }
                }
            }
        },
        "/sys/dpartment/update/{id}": {
            "post": {
                "description": "Update one dpartment record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Update one dpartment record",
                "parameters": [
                    {
                        "description": "Department Request Body for update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateResponse"
                        }
                    }
                }
            }
        },
        "/sys/dpartment/void/{id}": {
            "delete": {
                "description": "Delete one dpartment record by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Department"
                ],
                "summary": "Delete one dpartment record by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralUpdateInactiveResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DepartmentList": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "lists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Department"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.DepartmentPageDto": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GeneralCreateResponseBody": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateInactiveResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GeneralUpdateResponse"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GeneralUpdateResponse": {
            "type": "object",
            "properties": {
                "MatchedCount": {
                    "type": "integer"
                },
                "ModifiedCount": {
                    "type": "integer"
                },
                "UpsertedCount": {
                    "type": "integer"
                },
                "UpsertedID": {
                    "type": "string"
                }
            }
        },
...
        "models.Department": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deptCode": {
                    "type": "string"
                },
                "deptName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}